app-service/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
python-dotenv==1.0.1
remlaversionutilpy==0.1.1.dev1
Requests==2.31.0
